Alterando tabela de processos

+++ src/script_trackSecure.sql
@@ -75,7 +75,7 @@ CREATE TABLE Alerta (
 CREATE TABLE Processo (
 id INT PRIMARY KEY AUTO_INCREMENT,
 nome VARCHAR(45),
-usoMemoria INT,
+usoMemoria DECIMAL(5,2),
 fkServidor CHAR(17),
 FOREIGN KEY (fkServidor) REFERENCES Servidor (MacAddress)
 );
